Add js validation for acceptable file types (#524)
While the DropZone functionality validates the file extension,
this validation is not applied when selecting a file. The 
input[accept] property is not checked and if a user is able to
select a file with an unacceptable file type it is uploaded
to the server.

This patch prevents the above from happening.